COMP: Create Slicer4-EMSegment branch for current development.
Following discussion with Kilian Pohl, we agreed on creating
a dedicated branch to support the update of the module so that
it build and run against Slicer4.

It means that code pushed on [1] is expected for maintainance
of the EMSegment so that it build and run against Slicer3.

On the other hand, code pushed on [2] is non expected to be
backward compatible with Slicer3 and is done to ensure that
EMSegment build and run against Slicer4.

[1] http://svn.slicer.org/Slicer3/trunk/Modules/EMSegment
[2] http://svn.slicer.org/Slicer3/branches/Slicer4-EMSegment


For reference, the following command was used:

svn copy http://svn.slicer.org/Slicer3/trunk/Modules/EMSegment \
  http://svn.slicer.org/Slicer3/branches/Slicer4-EMSegment